pyproject.toml
[tool.poetry]
name = "numpy_vc_cupy"
version = "0.1.0"
description = ""
authors = ["Yuri Hartmann <yurihartmann0607@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
asyncer = "^0.0.2"
numpy = "^1.24.3"
cupy-cuda12x = "^12.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
